+++ code/engine.vc2008/xrGame/alife_storage_manager.cpp
@@ -67,11 +67,11 @@ void CALifeStorageManager::save	(LPCSTR save_name_no_check, bool update_name)
 		objects().save			(stream);
 		registry().save			(stream);
 
-		source_count			= stream.tell();
+		source_count			= (u32)stream.tell();
 		void					*source_data = stream.pointer();
 		dest_count				= rtc_csize(source_count);
 		dest_data				= xr_malloc(dest_count);
-		dest_count				= rtc_compress(dest_data,dest_count,source_data,source_count);
+		dest_count				= (u32)rtc_compress(dest_data,dest_count,source_data,source_count);
 	}
 
 	string_path					temp;
